Move getType to WiringUtils, make public
The `getType` method has general utility for other passes outside of
Wiring (I think!). This moves it to live under the WiringUtils object.

Additionally:
  * one unneeded parameter of a case object is ignored during pattern
    matching
  * documentation updates
